HW lighting refactor, add highlight effect to signs/targeting.
Moved lighting code to separate file to make working on it a bit
easier.
Added a new lighting flag, LIGHT_FLAG_HIGHLIGHT, which is used for
signs and the player's target to make them stand out more. Player's
target highlight color is the same as the outline (calculated 100%
client side, signs send lighting data).

Moved Raza Inn teleport coordinates/angle to face players at the
sign/NPC when they enter.
