+++ clientd3d/client.h
@@ -196,6 +196,7 @@ M59EXPORT void _cdecl dprintf(char *fmt,...);
 #include "d3dtypes.h"
 #include "d3dcache.h"
 #include "d3drender.h"
+#include "d3dlighting.h"
 #include "d3dparticle.h"
 #include "matrix.h"
 #include "d3ddriver.h"
