+++ clientd3d/clientd3d.vcxproj
@@ -170,6 +170,7 @@ $(SolutionDir)postbuild.bat
     <ClInclude Include="cursor.h" />
     <ClInclude Include="d3dcache.h" />
     <ClInclude Include="d3ddriver.h" />
+    <ClInclude Include="d3dlighting.h" />
     <ClInclude Include="d3dparticle.h" />
     <ClInclude Include="d3drender.h" />
     <ClInclude Include="d3dtypes.h" />
@@ -271,6 +272,7 @@ $(SolutionDir)postbuild.bat
     <ClCompile Include="cursor.c" />
     <ClCompile Include="d3dcache.c" />
     <ClCompile Include="d3ddriver.c" />
+    <ClCompile Include="d3dlighting.c" />
     <ClCompile Include="d3dparticle.c" />
     <ClCompile Include="d3drender.c" />
     <ClCompile Include="dialog.c" />
