+++ clientd3d/object3d.c
@@ -886,7 +886,7 @@ void DrawObjectDecorations(DrawnObject *object)
    if (r == NULL)
       return;
 
-   if (!(r->obj.flags & OF_DISPLAY_NAME || r->obj.flags & OF_SIGN)
+   if (!(r->obj.flags & OF_DISPLAY_NAME)
       || (r->obj.drawingtype == DRAWFX_INVISIBLE))
       return;
