+++ kod/include/blakston.khd
@@ -174,6 +174,7 @@
    LIGHT_FLAG_ON   = 0x0001
    LIGHT_FLAG_DYNAMIC = 0x0002
    LIGHT_FLAG_WAVERING = 0x0004
+   LIGHT_FLAG_HIGHLIGHT = 0x0008
 
    % Color constants for dynamic lights, in hexidecimal form.
    % Bright colors, 100% saturation
